loadGenerator/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2020.6.20        # via geventhttpclient-wheels, requests
chardet==3.0.4            # via requests
click==7.1.2              # via flask
configargparse==1.2.3     # via locustio
flask==1.1.2              # via locustio
gevent==20.4.0            # via geventhttpclient-wheels, locustio
geventhttpclient-wheels==1.3.1.dev2  # via locustio
greenlet==0.4.16          # via gevent
idna==2.10                # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
locustio==0.14.6          # via -r requirements.in
markupsafe==1.1.1         # via jinja2
msgpack==1.0.0            # via locustio
psutil==5.7.2             # via locustio
pyzmq==19.0.2             # via locustio
requests==2.24.0          # via locustio
six==1.15.0               # via geventhttpclient-wheels
urllib3==1.25.10          # via requests
werkzeug==1.0.1           # via flask
